simplify and abstract sort function
