+++ client/web/src/views/mocks/charts-content.ts
@@ -1,6 +1,6 @@
 import { LineChartContent } from 'sourcegraph'
 
-import { sortSeriesByName } from '../../insights/utils/sort-series-by-name'
+import { semanticSort } from '../../insights/utils/semantic-sort'
 
 export const LINE_CHART_CONTENT_MOCK: LineChartContent<any, string> = {
     chart: 'line',
@@ -29,7 +29,7 @@ export const LINE_CHART_CONTENT_MOCK: LineChartContent<any, string> = {
             name: 'B metric',
             stroke: 'var(--warning)',
         },
-    ].sort(sortSeriesByName),
+    ].sort((a, b) => semanticSort(a.name, b.name)),
     xAxis: {
         dataKey: 'x',
         scale: 'time',
@@ -69,7 +69,7 @@ export const LINE_CHART_TESTS_CASES_EXAMPLE: LineChartContent<any, string> = {
             name: 'React Testing Library',
             stroke: 'var(--red)',
         },
-    ].sort(sortSeriesByName),
+    ].sort((a, b) => semanticSort(a.name, b.name)),
     xAxis: {
         dataKey: 'x',
         scale: 'time',
@@ -107,7 +107,7 @@ export const LINE_CHART_WITH_MANY_LINES: LineChartContent<any, string> = {
         { dataKey: 'c', name: 'useTheme adoption', stroke: 'var(--blue)' },
         { dataKey: 'd', name: 'Class without CSS modules', stroke: 'var(--purple)' },
         { dataKey: 'f', name: 'Functional components without CSS modules', stroke: 'var(--green)' },
-    ].sort(sortSeriesByName),
+    ].sort((a, b) => semanticSort(a.name, b.name)),
     xAxis: {
         dataKey: 'x',
         scale: 'time',
@@ -263,7 +263,7 @@ export const LINE_CHART_WITH_HUGE_NUMBER_OF_LINES: LineChartContent<any, string>
         { dataKey: 'r', name: '1.23', stroke: 'var(--oc-grape-7)' },
         { dataKey: 's', name: '1.24', stroke: 'var(--oc-green-7)' },
         { dataKey: 't', name: '1.25', stroke: 'var(--oc-cyan-7)' },
-    ].sort(sortSeriesByName),
+    ].sort((a, b) => semanticSort(a.name, b.name)),
     xAxis: {
         dataKey: 'x',
         scale: 'time',
@@ -285,7 +285,7 @@ export const LINE_CHART_CONTENT_MOCK_EMPTY: LineChartContent<any, string> = {
             name: 'B metric',
             stroke: 'var(--warning)',
         },
-    ].sort(sortSeriesByName),
+    ].sort((a, b) => semanticSort(a.name, b.name)),
     xAxis: {
         dataKey: 'x',
         scale: 'time',
